+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added preview for decal shader graph (Only albedo, normal and emission)
 - Added exposure weight control for decal
 - Screen Space Directional Shadow under a define option. Activated for ray tracing 
+- Added multipass support for VR
+- Added XR SDK integration (multipass only)
 - Added a new abstraction for RendererList that will help transition to Render Graph and future RendererList API
 - Added Shader Graph samples for Hair, Fabric and Decal master nodes.